# app/backend/swaggerdoc.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manual Técnico do VolunteeringIPS",
        "version": "1.0.0",
        "title": "VolunteeringIPS Doc"
    },
    "basePath": "/api",
...
    "schemes": ["http"],
    "paths": {
        "/email/notify": {
            "post": {
                "summary": "Used to send an email to an user",
                "consumes": [
                    "application/json"
                ],
                "produces": ["application/json"],
                "tags": [
                    "Notifications"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "notification",
                        "description": "The notification to create",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Notification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email successfully sent"
                    },
                    "401": {
                        "description": "Email not sent due to unauthorization (invalid token)"
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "summary": "Get all users",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "You successfuly got the information"
                    }
                }
            },
            "post": {
                "summary": "Add an user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "addUser",
                        "description": "Insert all the information about the user",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Request successefully sent"
                    },
                    "400": {
                        "description": "Empty fields,  phone number or email already in DB, RGPD not accepted or error on saving data!"
                    },
                    "403": {
                        "description": "You are not an IPS community member"
                    }
                }
            }
        },
        "/user/setImage/{email}": {
            "post": {
                "summary": "Insert an image into an user",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Image to be inserted",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successefully sent"
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "403": {
                        "description": "Non existent email"
                    }
                }
            }
        },
        "/user/details/{id}": {
            "get": {
                "summary": "Get one user by is id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "You successfuly got the information"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user/delete/{id}": {
            "delete": {
                "summary": "Deletes one user by is id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user/update/{id}": {
            "put": {
                "summary": "Edit an user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "updateUser",
                        "description": "Updates all the information about the user",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/UpdateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successefully updated"
                    },
                    "400": {
                        "description": "Phone number or email already in DB!"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user/updatePassword/{id}": {
            "put": {
                "summary": "Updates an existing password",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "updatePassword",
                        "description": "First password is the actual one, the second is the new one",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/PasswordChange"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successefully updated"
                    },
                    "400": {
                        "description": "Old Password is not matching the existing one"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Notification": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "auth": {
                    "type": "string"
                },
                "action": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "params": {
                    "type": "object"
                }
            }
        },
...
        "User": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "isVolunteer": {
                    "type": "boolean"
                },
                "communityMemberType": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "schoolOrService": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "areasOfInterest": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "RGPD": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "isInactive": {
                    "type": "boolean"
                },
                "byAdmin": {
                    "type": "boolean"
                }
            }
        },
        "UpdateUser": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "isVolunteer": {
                    "type": "boolean"
                },
                "communityMemberType": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "schoolOrService": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "areasOfInterest": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "RGPD": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "isInactive": {
                    "type": "boolean"
                },
                "byAdmin": {
                    "type": "boolean"
                }
            }
        },
        "PasswordChange": {
            "type": "object",
            "properties": {
                "oldPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        }
    }
}